mev_arbitrage_bot/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
web3==6.20.3
python-dotenv==1.0.0
requests==2.31.0
eth-account==0.8.0
pydantic==1.10.8
